requirements.txt
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
torchmetrics==1.6.2
tensorboard==2.19.0

g2p-en==2.1.0
pypinyin==0.39.0
Unidecode==1.1.1
fairseq==0.12.2
librosa==0.9.1
matplotlib==3.7.2
nltk==3.9.1
numba==0.60.0
numpy==1.22.4
pandas==2.2.3
phonemizer==3.3.0
pip==23.3.1
protobuf==3.20.0
pyworld==0.3.4
PyYAML==6.0.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
tgt==1.4.4
tqdm==4.66.2
wheel==0.41.2
piq==0.8.0